+++ frontend/app/novels/novels-list/download-json-button.tsx
@@ -1,0 +1,33 @@
+import * as React from "react"
+import { Button } from "../../../components/ui/button"
+import { NovelEntry } from "./novel-types"
+import { entry_to_api } from "./novel-types"
+
+interface DownloadJsonButtonProps {
+  tableData: NovelEntry[]
+}
+
+export function DownloadJsonButton({ tableData }: DownloadJsonButtonProps) {
+  return (
+    <Button
+      variant="outline"
+      size="sm"
+      onClick={() => {
+        // export to api since it can be used in the backend
+        const novelTableData = tableData.map(entry_to_api);
+
+        const blob = new Blob([JSON.stringify(novelTableData)], { type: "application/json" });
+        const url = URL.createObjectURL(blob);
+        const link = document.createElement("a");
+        link.setAttribute("href", url);
+        link.setAttribute("download", `${Date.now()}.json`);
+        document.body.appendChild(link);
+        link.click();
+        link.remove();
+      }
+      }
+    >
+      Download JSON
+    </Button>
+  )
+}
